% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{fdc}
\alias{fdc}
\title{Title cross table}
\usage{
fdc(fdc.variable1, fdc.variable2)
}
\description{
Title cross table
}
\examples{
fdc(iris$Species, iris$Sepal.Length)
fhtml(fdc(iris$Species, iris$Sepal.Length))
}
