% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{fdiscrete}
\alias{fdiscrete}
\title{judge discrete variables}
\usage{
fdiscrete(d, t = 10)
}
\value{
judge discrete variables
}
\description{
judge discrete variables
}
\examples{
fdiscrete(iris$Sepal.Length)
fdiscrete(cars$dist)
}
