% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{fhtml}
\alias{fhtml}
\title{Title html view}
\usage{
fhtml(h, title = "Output", backgroundcolor = "black", textcolor = "white")
}
\description{
Title html view
}
\examples{
fhtml(summary(iris$Sepal.Length))
}
