% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{Fine.Gray.HRs}
\alias{Fine.Gray.HRs}
\title{Fine.Gray model for competing risk}
\usage{
Fine.Gray.HRs(FG.time, FG.status, FG.model, FG.n)
}
\arguments{
\item{FG.n}{}
}
\description{
Fine.Gray model for competing risk
}
