% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{fjcal}
\alias{fjcal}
\title{Title update Japanese calender}
\usage{
fjcal(J)
}
\description{
Title update Japanese calender
}
\examples{
fjcal(c("平 5", "令 1.12.12", "昭 64.1"))
fhtml(fjcal(c("平 5", "令 1.12.12", "昭 64.1")))
}
