% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{fn}
\alias{fn}
\title{Title}
\usage{
fn(Event, Exposure, fn.prop = FALSE, fn.test = FALSE)
}
\arguments{
\item{fn.exposure}{}
}
\description{
Title
}
\examples{
fn(iris$Petal.Length, iris$Species, T, F)
fn(sample(0:1, 150, replace = TRUE), iris$Species, T, T)
}
