% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{fpn}
\alias{fpn}
\title{Title}
\usage{
fpn(fpn.event, fpn.exposure, fpn.prop = FALSE, fpn.test = FALSE, outcome = "1")
}
\description{
Title
}
\examples{
fpn(mtcars$am , mtcars$gear)
fpn(mtcars$am , mtcars$gear, outcome = "0")
fpn(mtcars$am , mtcars$gear, T, T)
fhtml(fpn(mtcars$am , mtcars$gear))
}
