% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{ft}
\alias{ft}
\title{Title}
\usage{
ft(
  ft.data,
  ft.vars,
  ft.cats,
  ft.nvars = NULL,
  ft.group = NULL,
  ft.cdig = 1,
  ft.test = FALSE
)
}
\description{
Title
}
\examples{
ft(iris, colnames(iris),"Species","Petal.Length","Species")
fhtml(ft(iris, colnames(iris),"Species","Petal.Length","Species"))
}
