% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{update.J.cal}
\alias{update.J.cal}
\title{Title update Japanese calender}
\usage{
\method{update}{J.cal}(J)
}
\arguments{
\item{J}{}
}
\description{
Title update Japanese calender
}
\examples{
update.J.cal(c("平 5", "令 1.12.12"))
}
