% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{fe}
\alias{fe}
\title{Titlemodified summary}
\usage{
fe(fe.fit, fe.var = NULL, n = 2)
}
\arguments{
\item{n}{}
}
\description{
Titlemodified summary
}
\examples{
options(datadist =  datadist(iris)); iris$Species <-  sample(0:1, 150, replace = TRUE)

fe(lm(Sepal.Length ~ ., iris), 2:3)
fe(lm(Sepal.Length ~ ., iris), c("Petal.Length", "Sepal.Width"))

fe(ols(Sepal.Length ~ .,  subset(iris, select = -Species)), 2:3)
fe(ols(Sepal.Length ~ .,  subset(iris, select = -Species)), c("Petal.Length", "Sepal.Width"))

fe(summary(ols(Sepal.Length ~ .,  subset(iris, select = -Species)), Petal.Length = c(1,2)), 1:2)
fe(summary(ols(Sepal.Length ~ .,  subset(iris, select = -Species)), Petal.Length = c(1,2)), c("Petal.Length", "Sepal.Width"))

fe(lrm(Species ~.,  iris), 2:3)
fe(lrm(Species ~ .,  iris),c("Petal.Length", "Sepal.Width"))

fe(summary(lrm(Species ~ .,  iris), Sepal.Length = c(1,2),  Sepal.Width = c(1,2)), 2:3)
fe(summary(lrm(Species ~ .,  iris), Sepal.Length = c(1,2),  Sepal.Width = c(1,2)), c("Petal.Length", "Sepal.Width"))
}
