% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{fs}
\alias{fs}
\title{Title}
\usage{
fs(fs.varibale, fs.group = NA, fs.html = FALSE)
}
\arguments{
\item{fs.group}{}
}
\description{
Title
}
\examples{
 fs(iris$Sepal.Length, iris$Species)
 fs(iris$Sepal.Length)
}
