% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecodestack.R
\name{ft}
\alias{ft}
\title{Title}
\usage{
ft(
  T1.data,
  T1.vars,
  T1.cats,
  T1.nvars = NULL,
  T1.group = NULL,
  T1.cdig = 1,
  T1.test = FALSE
)
}
\arguments{
\item{T1.test}{}
}
\description{
Title
}
\examples{
ft (iris, colnames(iris),"Species","Petal.Length","Species")
}
